requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Werkzeug==2.2.3
SQLAlchemy==1.4.47
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0
pytz==2023.3